Crate thermal

A Finite Difference-based Thermal simulation module

It uses finite differences for marching forward in time and also for calculating the heat transfer through walls.

Modules

cavity

construction

environment

gas

glazing

heating_cooling

model

surface

zone

Constants

PI 🔒

SIGMA 🔒

The Stefan–Boltzmann constant (in $W/m^2K$), necessary for Radiation calculations

Type Definitions

Float 🔒