Module thermal::construction

Functions

build_thermal_network

Builds the necessary data for marching forward through time, solving the Ordinary Differential Equation that governs the heat transfer in walls.

calc_c_matrix

calc_k_matrix

Calculates the K matrix (i.e., the thermal network) for massive constructions

calc_n_total_nodes

This function calculates the number of nodes that result from the Construction’s discretization.

calc_r_back

calc_r_front

discretize_construction

Given a Maximum element thickness ($\Delta x_{max}$) and a minimum timestep ($\Delta t_{min}$), this function will find an arguibly good (i.e., stable and accurate) combination of $\Delta t$ and number of elements in each layer of the construction.

get_first_and_last_massive_elements

In a discretization scheme, this function finds the first and the last massive layers… the no-mass layers before and after the first and last massive ones, respectively, are just bulked with the values of RSi and RSo.