Function thermal::construction::build_thermal_network

pub fn build_thermal_network(
    construction: &Rc<Construction>, 
    first_massive: usize, 
    last_massive: usize, 
    dt: f64, 
    all_nodes: usize, 
    n_elements: &[usize], 
    r_front: f64, 
    r_back: f64
) -> Result<impl Fn(&Matrix, f64, f64, f64, f64, f64, f64, f64, f64) -> Matrix, String>

Builds the necessary data for marching forward through time, solving the Ordinary Differential Equation that governs the heat transfer in walls.

The equation to solve is the following:

\overline{C}  \dot{T} - \overline{K}  T = q

Where $\overline{C}$ and $\overline{K}$ are matrices representing the thermal mass of each node and the thermal network, respectively; and where $T$ and $q$ are vectors representing the Temperature and the heat “flow into” each node, respectively. $\overline{C}$ and $\overline{K}$ are build based on the finite difference method.

This model uses a 4th order Runga-Kutte (a.k.a., RK4) to march through time. In order to do this, it is convenient to write the equation to solve as follows:

\dot{T}  = f(t, T)

Where

f(t,T) = \overline{C}^{-1} \overline{K}  T + \overline{C}^{-1} q

Note that—unless some layer of the Surface is generating heat—all the elements of $q$ are Zero exept the first one and the last one, which are $\frac{T_{in}}{R_{si}}$ and $\frac{T_{out}}{R_{so}}$, respectively.

Then, the 4th order Runge-Kutta method allows marching forward through time as follows:

 T_{i+1} = T_i + \frac{k_1 + 2k_2 + 2k_3 + k_4}{6}

Where $k_1$, $k_2$, $k_3$ and $k_4$ can be calculated based on the timestep $\Delta t$ as follows:

$k_1 = \Delta t \times f(t,T)$

$k_2 = \Delta t \times f(t+\frac{\Delta t}{2}, T+\frac{k_1}{2})$

$k_3 = \Delta t \times f(t+\frac{\Delta t}{2}, T+\frac{k_2}{2})$

$k_4 = \Delta t \times f(t+\delta t, T+k_3 )$

So, what this method does is to calculate what is needed in order to return a closure that calculates $\Delta t \times f(t,T)$; that is to say:

return = \Delta t \times f(t,T) = \Delta t \times \overline{C}^{-1} \overline{K}  T + \Delta t \times \overline{C}^{-1} q

It is worth mentioning that, due to efficiency reasons, the following variables are defined within the code:

$\overline{K}^{\star} = \Delta t \times \overline{C}^{-1}\overline{K}$

$\overline{C}^{\star} = \Delta t \times \overline{C}^{-1}$