Function thermal::construction::calc_k_matrix

fn calc_k_matrix(
    c: &Rc<Construction>, 
    first_massive: usize, 
    last_massive: usize, 
    n_elements: &[usize], 
    all_nodes: usize
) -> Matrix

Calculates the K matrix (i.e., the thermal network) for massive constructions

Constructions are assumed to be a sandwich where zero or more massive layers are located between two non-mass layers. These non-mass layers will always include the interior and exterior film convections coefficients, respectively (which is why they are called full_rs_front and full_rs_back, respectively). Additionally, they can also include some lightweight insulation or any other material of negligible thermal mass.

This matrix is constructed based ona discretization of the layers of the construction; that is, each layer is subdivided into n elements all of equal thickness. One node is placed at the beginning and end of each elements. Each element can be represented by the 2x2 matrix

\overline{K}=\begin{bmatrix} -1/R & 1/R \\
1/R & -1/R
\end{bmatrix}   ;   R=\frac{thickness}{\lambda}

Hence—ignoring external inputs—the K matrix for a construction subdivided into 3 elements (i.e., 4 nodes) can be written as follows:

\overline{K}=\begin{bmatrix}
-1/R_{1\rightarrow2} & 1/R_{1\rightarrow2} & 0 & 0 \\
1/R_{1\rightarrow2} & -1/R_{1\rightarrow2} - 1/R_{2\rightarrow3} & 1/R_{2\rightarrow3} & 0 \\
0 & 1/R_{2\rightarrow3} & -1/R_{2\rightarrow3} - 1/R_{3\rightarrow4} & 1/R_{3\rightarrow4} \\
0 & 0 & 1/R_{3\rightarrow4} & -1/R_{3\rightarrow4} \\
\end{bmatrix}   

Now, these nodes are also connected to an interior and an exterior temperatures through all the layers that do not have any thermal mass both in the interior and exterior (i.e., $R_{si,full}$ and $R_{so,full}$, respectively). This means that the Matrix $\overline{K}$ needs to become:

\overline{K}=\begin{bmatrix}
-1/R_{1\rightarrow2} - 1/R_{si,full} & 1/R_{1\rightarrow2} & 0 & 0 \\
1/R_{1\rightarrow2} & -1/R_{1\rightarrow2} - 1/R_{2\rightarrow3} & 1/R_{2\rightarrow3} & 0 \\
0 & 1/R_{2\rightarrow3} & -1/R_{2\rightarrow3} - 1/R_{3\rightarrow4} & 1/R_{3\rightarrow4} \\
0 & 0 & 1/R_{3\rightarrow4} & -1/R_{3\rightarrow4}- 1/R_{so,full} \\
\end{bmatrix}   

This method returns such a matrix, without the $R_{si, full}$ and $R_{so, full}$. They need to be added when marching (because these values change over time).