Function thermal::construction::discretize_construction

pub fn discretize_construction(
    construction: &Rc<Construction>, 
    model_dt: f64, 
    max_dx: f64, 
    min_dt: f64
) -> (usize, Vec<usize>)

Given a Maximum element thickness ($\Delta x_{max}$) and a minimum timestep ($\Delta t_{min}$), this function will find an arguibly good (i.e., stable and accurate) combination of $\Delta t$ and number of elements in each layer of the construction.

This function recursively increases the model’s timestep subdivisions (n) in order to reduce $\Delta t$ to numbers that respect the restrictions of (1) stability, (2) $\Delta x_{max}$, and (3) $\Delta t_{min}$. In other words, it searches (by testing $\Delta t_{model}/1$, $\Delta t_{model}/2$, $\Delta t_{model}/3$, … $\Delta t_{model}/n$) for the minimum n that respects this restrictions

The math behind it

The first thing to know is that the walls in this module march through time using a 4th order Runga-Kutte (a.k.a., RK4). The second thing to know is that the RK4 method is more stable than the Euler method, and thus the restrictions of stability for the Euler method can be considered to be a conservative restriction for the RK4. Hence, this function uses the Euler method restrictions.

Now, as explained in the build_thermal_network documentation, we are solving the following equation:

\dot{T} = \overline{C}^{-1} \overline{K}  T + \overline{C}^{-1} q

And thus the stability of the numerical method will depend on the matrix:

\overline{K}^{\star} =\Delta t \overline{C}^{-1} \overline{K}

Specifically, we don’t want any of its eigenvalues $\xi_1, \xi_2,\xi_3, ...$ to be outside of the Euler method’s stability region. Since this matrix has onle Real eigenvalues, this is equivalent to saying:

-2 < \xi_i < 0 ; \forall i

However, finding the eigenvalues for $\overline{K}^{\star}$ is far from trivial. So there is yet another heuristic I am using: I am treating the case of a wall with 1 layer that is subdivided into a single element as the limit case. I am not sure if this is correct, but most of the instabilities I identified through Trial and Error corresponded to this case.

For this limit case:

$R = \frac{\Delta x}{\lambda}$

$C = \rho c_p \Delta x$

thus the value of $\overline{K}^{\star}$ is:

\overline{K}^{\star}=\begin{bmatrix} 
-\frac{\Delta t}{C\times R} - \frac{\Delta t}{C\times R_s} & \frac{\Delta t}{C\times R} \\
 \frac{\Delta t}{C\times R} & -\frac{\Delta t}{C\times R} - \frac{\Delta t}{C\times R_s}\\
\end{bmatrix}   

Note that, in that equation, $R_{si} = R_{so}$. The reason for this is that this method does not know the real values of $R_{si}$ and $R_{so}$, so it simply using a placeholder value low enough to cover most general cases (0.05).

Then, it can be found that the eigenvaues of this case—which we are treating as the limit case—are:

\xi_1 = -\frac{\Delta t} { R_s \rho c_p \Delta x }

\xi_2 = \xi_1 - 2 \frac{\Delta t \lambda}{ \rho  c_p  {\Delta x}^2}

Both these values are always negative, so we don’t need to worry about $\xi_i < 0 $. Also, it can be noticed that $\xi_2 < \xi_1$, meaning that what we need to comply with for Euler’s stability criteria is:

-\frac{\Delta t} { R_s \rho c_p \Delta x } - 2 \frac{\Delta t \lambda}{ \rho  c_p  {\Delta x}^2} < -2

Which means that the chosen $\Delta x$ must be greater than the (apparently only) positive solution to equation:

0 = 2 {\Delta x}^2 - \left( \frac{\Delta t}{\rho c_p R_s} \right) \Delta x - \frac{2 \Delta t \lambda}{\rho c_p} 

So, this method will identify a combination of $\Delta t$ and $\Delta x$ that allows complying with this