Struct thermal::construction::Discretization

pub struct Discretization {
    pub segments: Vec<(Float, UValue)>,
    pub tstep_subdivision: usize,
}

Represents the discretization of a [Construction] for heat transfer calculation purposes.

Note

This object contains all the Cavity objects in it, which contain information not only about their thickness but also their orientation. This means that one Discretization should exist per Surface, not just by Construction.

Fields
segments: Vec<(Float, UValue)>

Contains the node’s mass and the UValue of each segment

tstep_subdivision: usize

Contains the minimum number of timesteps per model timestep that this discretization requires to ensure numerical stability and accuracy.

While the caller model—e.g., a Multiphysics Simulation model—may attempt to simulate at a certain timestep, some of the constructions in it may require smaller timesteps to ensure numerical stability and accuracy. This means that—on each timestep in the caller model—the thermal model needs to perform time_subdivision sub-timesteps.

Implementations

impl Discretization

pub fn new(
construction: &Rc<Construction>,
model_dt: Float,
max_dx: Float,
min_dt: Float,
height: Float,
angle: Float
) -> Result<Self, String>

Creates a new Discretization.

It first calculates the tstep_subdivision and the number of elements on each layer of Construction by calling discretize_construction(); and then builds the Discretization by calling build().

fn chunk_segments(&self, indexes: &[usize]) -> Vec<(usize, usize)>

pub fn get_chunks(&self) -> (Vec<(usize, usize)>, Vec<(usize, usize)>)

fn build(
construction: &Rc<Construction>,
tstep_subdivision: usize,
n_elements: &[usize],
height: Float,
angle: Float
) -> Result<Self, String>

Creates the segments of the Discretization.

pub fn r_value(&self) -> Float

Calculates the R value of the whole system

Panics

Panics if the calculated R value is Zero (i.e., if there are no layers or something like that)

fn discretize_construction(
construction: &Rc<Construction>,
model_dt: Float,
max_dx: Float,
min_dt: Float
) -> (usize, Vec<usize>)

Given a Maximum element thickness ($\Delta x_{max}$) and a minimum timestep ($\Delta t_{min}$), this function will find an arguibly good (i.e., stable and accurate) combination of $\Delta t$ and number of elements in each layer of the construction.

This function recursively increases the model’s timestep subdivisions (n) in order to reduce $\Delta t$ to numbers that respect the restrictions of (1) stability, (2) $\Delta x_{max}$, and (3) $\Delta t_{min}$. In other words, it searches (by testing $\Delta t_{model}/1$, $\Delta t_{model}/2$, $\Delta t_{model}/3$, … $\Delta t_{model}/n$) for the minimum n that respects this restrictions

The math behind it

The first thing to know is that the walls in this module march through time using a 4th order Runga-Kutte (a.k.a., RK4). The second thing to know is that the RK4 method is more stable than the Euler method, and thus the restrictions of stability for the Euler method can be considered to be a conservative restriction for the RK4. Hence, this function uses the Euler method restrictions.

We are solving the following equation:

\dot{T} = \overline{C}^{-1} \overline{K}  T + \overline{C}^{-1} q

And thus the stability of the numerical method will depend on the matrix:

\overline{K}^{\star} =\Delta t \overline{C}^{-1} \overline{K}

Specifically, we don’t want any of its eigenvalues $\xi_1, \xi_2,\xi_3, ...$ to be outside of the Euler method’s stability region. Since this matrix has onle Real eigenvalues, this is equivalent to saying:

-2 < \xi_i < 0 ; \forall i

However, finding the eigenvalues for $\overline{K}^{\star}$ is far from trivial. So there is yet another heuristic I am using: I am treating the case of a wall with 1 layer that is subdivided into a single element as the limit case. I am not sure if this is correct, but most of the instabilities I identified through Trial and Error corresponded to this case.

For this limit case:

$R = \frac{\Delta x}{\lambda}$

$C = \rho c_p \Delta x$

thus the value of $\overline{K}^{\star}$ is:

\overline{K}^{\star}=\begin{bmatrix}
-\frac{\Delta t}{C\times R} - \frac{\Delta t}{C\times R_s} & \frac{\Delta t}{C\times R} \\
 \frac{\Delta t}{C\times R} & -\frac{\Delta t}{C\times R} - \frac{\Delta t}{C\times R_s}\\
\end{bmatrix}   

Note that, in that equation, $R_{si} = R_{so}$. The reason for this is that this method does not know the real values of $R_{si}$ and $R_{so}$, so it simply using a placeholder value low enough to cover most general cases (0.05).

Then, it can be found that the eigenvaues of this case—which we are treating as the limit case—are:

\xi_1 = -\frac{\Delta t} { R_s \rho c_p \Delta x }

\xi_2 = \xi_1 - 2 \frac{\Delta t \lambda}{ \rho  c_p  {\Delta x}^2}

Both these values are always negative, so we don’t need to worry about $\xi_i < 0 $. Also, it can be noticed that $\xi_2 < \xi_1$, meaning that what we need to comply with for Euler’s stability criteria is:

-\frac{\Delta t} { R_s \rho c_p \Delta x } - 2 \frac{\Delta t \lambda}{ \rho  c_p  {\Delta x}^2} < -2

Which means that the chosen $\Delta x$ must be greater than the (apparently only) positive solution to equation:

0 = 2 {\Delta x}^2 - \left( \frac{\Delta t}{\rho c_p R_s} \right) \Delta x - \frac{2 \Delta t \lambda}{\rho c_p}

So, this method will identify a combination of $\Delta t$ and $\Delta x$ that allows complying with this

All that said, the value for $\Delta t$ actually used by the final model is actually half of what these equations use. This is because what we are using is a heuristic and I want to be safe… ish

pub fn get_k_q(
&self,
ini: usize,
fin: usize,
temperatures: &Matrix,
front_env: &Environment,
front_emmisivity: Float,
front_hs: Float,
back_env: &Environment,
back_emmisivity: Float,
back_hs: Float
) -> (Matrix, Matrix)

Produces $\overline{K}$ and $\vec{q}$ (as in the equation $\overline{C} \dot{\vec{T}} = \overline{K} \vec{T} + \vec{q}$), allowing to update the node temperatures in a surface. This method returns the matrix $\overline{k}$ representing the thermal network, and the vector $\vec{q}$), accounting for the heat flows induced by the environment.

How it works

This matrix is constructed based ona discretization of the layers of the construction; that is, each layer is subdivided into n elements all of equal thickness. One node is placed at the beginning and end of each elements. Each element can be represented by the 2x2 matrix

\overline{K}=\begin{bmatrix} -U & U \\
U & -U
\end{bmatrix}   

Then—ignoring all external inputs (i.e., a system disconnected from the environment—the K matrix for a construction subdivided into 3 elements (i.e., 4 nodes) can be written as follows:

\overline{K}=\begin{bmatrix}
-U_{1\rightarrow2} & U_{1\rightarrow2} & 0 & 0 \\
U_{1\rightarrow2} & -U_{1\rightarrow2} - U_{2\rightarrow3} & U_{2\rightarrow3} & 0 \\
0 & U_{2\rightarrow3} & -U_{2\rightarrow3} - U_{3\rightarrow4} & U_{3\rightarrow4} \\
0 & 0 & U_{3\rightarrow4} & -U_{3\rightarrow4} \\
\end{bmatrix}   

Now, these nodes are also connected to a front and back border conditions. This means that the Matrix $\overline{K}$ needs to become:

\overline{K}=\begin{bmatrix}
-U_{1\rightarrow2} - h_{so} & U_{1\rightarrow2} & 0 & 0 \\
U_{1\rightarrow2} & -U_{1\rightarrow2} - U_{2\rightarrow3} & U_{2\rightarrow3} & 0 \\
0 & U_{2\rightarrow3} & -U_{2\rightarrow3} - U_{3\rightarrow4} & U_{3\rightarrow4} \\
0 & 0 & U_{3\rightarrow4} & -U_{3\rightarrow4}- h_{si} \\
\end{bmatrix}   

Additionally, the vector $\vec{q}$ will have to account for the border conditions. How this is done depends on the border condition. If the border leads to a zone, a value of $ T_{env} h_s + E_{ir}\epsilon_s - \epsilon_s \sigma {T_s}^4$ should be added to $\vec{q}$. Note that $T_{env}$ is the temparture of the air in the environment, $h_s$ is the convection coefficient, $E_{ir}$ is the incident infrared radiation and $\epsilon_s$ is the emmisivity of the surface. On the contrary, if the border condition is a cavity, then a value of $T_{pane} U_{cavity}$ should be added. $T_{pane}$ is the temperature of the surface before or after.