Module thermal::gas

Structs

Gas

A structure containing the data that will describe the thermal behaviour of a gas.

Functions

in_kelvin 🔒

Transforms C into K

nu_0_60 🔒

Calculates the Nusselt number for cavities tilted between $0^o$ and $60^o$

nu_60 🔒

Calculates the Nusselt number for cavities tilted $60^o$

nu_60_90 🔒

Calculates the Nusselt number for cavities tilted between $60^o$ and $90^o$

nu_90 🔒

Calculates the Nusselt number for cavities tilted $90^o$

nu_90_180 🔒

Calculates the Nusselt number for cavities tilted between $90^o$ and $180^o$

nusselt 🔒

Calculates the Nusselt of a cavity number based on the Raleigh number ra, the tilt of the cavity gamma and the aspect ratio a_gi (i.e., $a_{gi}$).