Struct thermal::gas::Gas

pub struct Gas {
    thermal_conductivity: Polynomial,
    dynamic_viscosity: Polynomial,
    heat_capacity: Polynomial,
    mass: Float,
}

A structure containing the data that will describe the thermal behaviour of a gas.

Fields
thermal_conductivity: Polynomial

The thermal conductivity (${W}/{m.K}$) as a function of the temperature (in $K$)

dynamic_viscosity: Polynomial

The dynamic viscosity ( ${N.s}/{m^2}$) as a function of the temperature (in $K$)

heat_capacity: Polynomial

The specific heat capacity (${J}/{kg.K}$) as a function of the temperature (in $K$)

mass: Float

THe Molecular Mass (${kg}/{Mol}$)

Implementations

impl Gas

fn raleigh(&self, t_front: Float, t_back: Float, thickness: Float) -> Float

Calculates the Raleigh number of a Gas cavity based on its thickness and its temperatures t_front and t_back (note that, for this particular function, these values are interchangeable)

Source: Equation 40 of ISO15099/2003

pub fn cavity_convection(
&self,
height: Float,
thickness: Float,
gamma: Float,
t_front: Float,
t_back: Float
) -> Float

Calculates the convective heat transfer coefficient within a gas-filled cavity based on its tilt $\gamma$ (in radians), its outside temperature t_out and its interior temperature t_in.

Adapted from Section 5.3.3.1 of ISO15099/2003:

$\gamma = 0^o$: is horizontal glazing

$\gamma = 90^o$: is vertical glazing

Note:

The standard ISO15099/2003 states the following:

“This categorization, as a function of $\gamma$, is based on the assumption that the cavity is heated from the internal side (i.e., $T_{ft,i} > T_{b,i-1}$). If the reverse is true it is necessary to seek the appropriate correlation on the basis of the complement of the tilt angle, $180^o - \gamma$, instead of $\gamma$ when the calculation is carried out”.

This conversion is handleded automatically by this function based on the inputs given to t_front and t_back

pub fn thermal_conductivity(&self, temp: Float) -> Float

Derives the Thermal Conductivity at a certain Temperature (in $K$)

pub fn dynamic_viscosity(&self, temp: Float) -> Float

Derives the Dynamic Viscosity at a certain Temperature (in $K$)

pub fn heat_capacity(&self, temp: Float) -> Float

Derives the Soecific Heat Capacity at a certain Temperature (in $K$)

pub fn mass(&self) -> Float

Retreives the Molecular Mass

pub fn density(&self, temp: Float) -> Float

Derives the density based on the temperature (in $K$)

pub fn air() -> Self

Returns a gas with the properties of Air

pub fn argon() -> Self

Returns a gas with the properties of argon

pub fn krypton() -> Self

Returns a gas with the properties of krypton

pub fn xenon() -> Self

Returns a gas with the properties of xenon