Enum thermal::surface::ThermalSurface

pub enum ThermalSurface {
    Fenestration(Rc<Fenestration>, ThermalSurfaceData),
    Surface(Rc<Surface>, ThermalSurfaceData),
}

Variants

Fenestration(Rc<Fenestration>, ThermalSurfaceData)

Tuple Fields of Fenestration

0: Rc<Fenestration>1: ThermalSurfaceData

Surface(Rc<Surface>, ThermalSurfaceData)

Tuple Fields of Surface

0: Rc<Surface>1: ThermalSurfaceData

Implementations

impl ThermalSurface

pub fn new_surface(
state: &mut SimulationStateHeader,
surface: &Rc<Surface>,
dt: f64,
n_elements: &[usize]
) -> Result<Self, String>

pub fn new_fenestration(
state: &mut SimulationStateHeader,
fenestration: &Rc<Fenestration>,
dt: f64,
n_elements: &[usize]
) -> Result<Self, String>

fn data(&self) -> &ThermalSurfaceData

Borrows the ThermalSurfaceData of this surface

fn mut_data(&mut self) -> &mut ThermalSurfaceData

Borrows a mutable version of ThermalSurfaceData of this surface

pub fn area(&self) -> f64

Gets the area of the surface

pub fn rs_front(&self) -> f64

Gets the rs_front

pub fn rs_back(&self) -> f64

Gets the rs_back

pub fn set_front_boundary(&mut self, b: Boundary)

pub fn set_back_boundary(&mut self, b: Boundary)

pub fn front_boundary(&self) -> &Option<Boundary>

pub fn back_boundary(&self) -> &Option<Boundary>

pub fn is_massive(&self) -> bool

Checks whether a wall has thermal mass

pub fn front_temperature(&self, state: &SimulationState) -> f64

pub fn back_temperature(&self, state: &SimulationState) -> f64

pub fn set_node_temperatures(
&self,
state: &mut SimulationState,
matrix: &Matrix
)

pub fn get_node_temperatures(&self, state: &SimulationState) -> Matrix

Retrieves the state of the Surface as a Matrix object.

fn calc_heat_flow(
&self,
state: &SimulationState,
t_front: f64,
t_back: f64
) -> (f64, f64)

Calculates the heat flow out of the layer, based on the inside and outside temperatures

pub fn march(
&self,
state: &mut SimulationState,
t_front: f64,
t_back: f64,
_dt: f64
) -> (f64, f64)

Marches one timestep. Returns front and back heat flow