Struct thermal::surface::ThermalSurfaceData

pub struct ThermalSurfaceData<T: SurfaceTrait> {

    pub parent: Rc<T>,
    pub discretization: Discretization,
    pub front_boundary: Option<Boundary>,
    pub back_boundary: Option<Boundary>,
    pub front_emmisivity: Float,
    pub back_emmisivity: Float,
    pub front_solar_absorbtance: Float,
    pub back_solar_absorbtance: Float,
    pub area: Float,
    pub massive_chunks: Vec<(usize, usize)>,
    pub nomass_chunks: Vec<(usize, usize)>,
    pub front_alphas: Matrix,
    pub back_alphas: Matrix,

}

This is a Surface from the point of view of our thermal solver. Since this module only calculate heat transfer (and not short-wave solar radiation, e.g., light), both simple_model::Fenestration and simple_model::Surface are treated in the same way.

Fields
parent: Rc<T>discretization: Discretization

The Discretization that represents this ThermalSurfaceData

front_boundary: Option<Boundary>

The location of the front boundary zone in the Zones array of the Thermal Model

back_boundary: Option<Boundary>

The location of the back boundary zone in the Zones array of the Thermal Model

front_emmisivity: Float

The thermal absorbtance on the front side (from 0 to 1)

back_emmisivity: Float

The thermal absorbtance on the back side (from 0 to 1)

front_solar_absorbtance: Float

The solar absorbtance on the front side (from 0 to 1)

back_solar_absorbtance: Float

The solar absorbtance on the back side (from 0 to 1)

area: Float

The area of the Surface

massive_chunks: Vec<(usize, usize)>

The chunks of nodes that have mass

nomass_chunks: Vec<(usize, usize)>

The chunks of nodes that have nomass

front_alphas: Matrix

The absorbtances of each node in the system, proportional to the front incident radiation (i.e., they do not add up to 1.0)

back_alphas: Matrix

The absorbtances of each node in the system, proportional to the back incident radiation (i.e., they do not add up to 1.0)

Implementations

impl<T: SurfaceTrait> ThermalSurfaceData<T>

pub fn new(
state: &mut SimulationStateHeader,
ref_surface_index: usize,
parent: &Rc<T>,
area: Float,
construction: &Rc<Construction>,
discretization: Discretization
) -> Result<ThermalSurfaceData<T>, String>

pub fn set_front_boundary(&mut self, b: Boundary)

pub fn set_back_boundary(&mut self, b: Boundary)

pub fn march(
&self,
state: &mut SimulationState,
t_front: Float,
t_back: Float,
dt: Float
) -> (Float, Float)

Marches one timestep. Returns front and back heat flow