Struct thermal::model::ThermalModel

pub struct ThermalModel {
    pub zones: Vec<ThermalZone>,
    pub surfaces: Vec<ThermalSurface>,
    pub fenestrations: Vec<ThermalFenestration>,
    pub dt_subdivisions: usize,
    pub dt: Float,
}

Fields
zones: Vec<ThermalZone>

All the Thermal Zones in the model

surfaces: Vec<ThermalSurface>

All the surfaces in the model

fenestrations: Vec<ThermalFenestration>

All the Fenestrations in the model

dt_subdivisions: usize

The number of steps that this model needs to take in order to advance one step of the main simulation.

dt: Float

The model’s dt (i.e., main_dt / self.dt_subdivisions)

Implementations

impl ThermalModel

pub fn dt_subdivisions(&self) -> usize

Retrieves the dt_subdivisions (i.e. the number of substimesteps per timestep of this model)

pub fn get_thermal_zone(&self, index: usize) -> Result<&ThermalZone, String>

Retrieves a ThermalZone

fn calculate_zones_abc(
&self,
model: &SimpleModel,
state: &SimulationState
) -> (Vec<Float>, Vec<Float>, Vec<Float>)

This estimation assumes nothing changes during this time. This is self evidently wrong, as we know that, for example, the surface temperatures will change together with the zone air temperature. However, in short periods of time this can actually work.

Everything starts from the following equation, representing a heat balance over the air and the contents of the Thermal zone.

C_{zone}\frac{dT(t)}{dt} = \displaystyle\sum_{i=loads}{Q_i} + \displaystyle\sum_{i=surf.}{h_iA_i(T_i-T)}+\displaystyle\sum_{i=otherzones}{\dot{m_i}C_p(T_i-T)}+\dot{m}_{inf}C_p(T_{out}-T)+\dot{m}_{supplied}C_p(T_{sup}-T)

Which can be expanded into the following

C_{zone}\frac{dT(t)}{dt} = A - B T

Where $A$ and $B$ are constant terms (at least they can be assumed to be during this brief period of time).

A = \displaystyle\sum_{i=loads}{Q_i} + \displaystyle\sum_{i=surf.}{h_iA_i T_i}+\displaystyle\sum_{i=otherzones}{\dot{m_i}C_pT_i}+\dot{m}_{inf}C_pT_{out}+\dot{m}_{supplied}C_pT_{sup}

B= \displaystyle\sum_{i=surf.}{h_iA_i}+\displaystyle\sum_{i=otherzones}{\dot{m_i}C_p}+\dot{m}_{inf}C_p +\dot{m}_{supplied}C_p

And so, (solving the differential equation) the Temperature $T$ at a time $t$ into the future can be estimated based on the current Temperature of the zone ($T_{current}$) and the following equation:

 T(t) = \frac{A}{B} + \left( T_{current} - \frac{A}{B} \right)e^{-\frac{B}{C_{zone}}t}

And the average temperature during the same periood is:

\frac{\displaystyle\int_{0}^t{T(t)dt}}{t} = \frac{A}{B}+\frac{C_{zone}\left(T_{current}-\frac{A}{B}\right)}{Bt}\left(1-e^{-\frac{Bt}{C_{zone}}} \right)

fn get_current_zones_temperatures(&self, state: &SimulationState) -> Vec<Float>

Retrieves a vector of the current temperatures of all the Zones as registered in the Simulation State

fn estimate_zones_mean_future_temperatures(
&self,
t_current: &[Float],
a: &[Float],
b: &[Float],
c: &[Float],
future_time: Float
) -> Vec<Float>

Uses an analytical solution to estimate an average temperature for each Zone for the near future. Uses the coefficients $A$, $B$ and $C$ calculated by calculate_zones_abc and the Zones’ current temperatures t_current as calculated by get_current_temperatures.

fn estimate_zones_future_temperatures(
&self,
t_current: &[Float],
a: &[Float],
b: &[Float],
c: &[Float],
future_time: Float
) -> Vec<Float>

Uses an analytical solution to estimate the future Zones temperature for the near future. Uses the coefficients $A$, $B$ and $C$ calculated by calculate_zones_abc and the Zones’ current temperatures t_current as calculated by get_current_temperatures.